Finishing touch

+++ ekantipur/Scrapper.py
@@ -1,13 +1,15 @@
 # -*- coding: utf-8 -*-
 
-from bs4 import BeautifulSoup
-import urllib.request
-from urllib.parse import urlparse
 import re
 import json
 import sys
 import os
-import requests
+import argparse
+
+import urllib.request
+from bs4 import BeautifulSoup
+from urllib.parse import urlparse
+
 
 class ReadOnlyClass(type):
     def __setattr__(self, name, value):
@@ -17,7 +19,7 @@ class ReadOnlyClass(type):
 class Scrapper:
     __metaclass__ = ReadOnlyClass
 
-    def __init__(self, news_link='', source=''):
+    def __init__(self, news_link='', source='', given_date=''):
         self.NEWS_LINK = news_link
         self.SOURCE = source
         self.dump = {
@@ -26,15 +28,26 @@ class Scrapper:
             'category' : {
             }
         }
+        self.given_date = given_date
+        self.published_date = ''
         
         if not os.path.exists(self.SOURCE):
             os.mkdir(self.SOURCE)
-
         
-    def saveJson(self, filename=None):
-        fname = os.path.join(self.SOURCE, filename+'.json')
-        with open(fname, 'w', encoding='utf-8') as f:
-            json.dump(self.dump, f, ensure_ascii=False, indent=4)
+    def saveJson(self, directory=None, input_file=None):
+        save_path = os.path.join(self.SOURCE, directory)
+        
+        if not os.path.exists(save_path):
+            os.mkdir(save_path)
+            
+        filename = input_file + '.json'
+        fname = os.path.join(save_path, filename)
+        
+        # Only write when file not present
+        if not os.path.exists(fname):
+            print("Saving file : ", fname)
+            with open(fname, 'w', encoding='utf-8') as f:
+                json.dump(self.dump, f, ensure_ascii=False, indent=4)
             
 
     def extractContent(self):
@@ -71,6 +84,8 @@ class Scrapper:
             # Get category name in English
             category_eng = url.split("/")[-1]
             
+            url += '/'+self.given_date
+            
             r = urllib.request.urlopen(url).read()
             soup = BeautifulSoup(r, 'html.parser')
 
@@ -81,84 +96,114 @@ class Scrapper:
                     if 'html' in data.find('a').get('href'):
                         headlineList.append((category, data.find('a').get('href')))
                 
-            result = self.newsContents(headlineList)
-            if len(result) > 0:
-                self.dump['category'] = result 
-                self.saveJson(category_eng)
-                
-                
+            self.newsContents(headlineList)
+
 
     # retrieve the body for each headline
     def newsContents(self, headlineList):
         
-        news_dump = {}
-        for index, (category, half_link) in enumerate(headlineList):
+        # Get the date wise dumps
+        date_dump = {}
+        for cat, each_link in headlineList:
+            curr_link = each_link.split('/')
+            current_date = curr_link[3]+curr_link[4]+curr_link[2] 
             
-            print ("*************** Category : {}, #: {} ********************".format(category, (index + 1)))
-            url = ''
+            if current_date in date_dump:
+                date_dump[current_date].append(each_link)
+            else:
+                date_dump[current_date] = [each_link]
             
-            # [1:] because of extra / in half_link
-            url = url.join((self.NEWS_LINK, half_link[1:]))
-            print("Link :", url)
-            
-            r = urllib.request.urlopen(url).read()
-            soup = BeautifulSoup(r, 'html.parser')
-            
-            title_source = soup.find_all('div', {'class': ['article-header']})
-            if title_source[0].find({'h1', 'h2'}) is not None:
-                news_title = title_source[0].find({'h1', 'h2'}).text
-    
-            author = soup.find('span',{'class': ['author']}).text
-            nep_date = soup.find('time',{'style': ['display:inline-block']}).text           
-            
-            # Get only the content from 
-            # normal article of main page of each link
-            body_content = soup.find('article', {'class': ['normal']})
-
-            news_body=' '
+        # Get content date-wise
+        # And store accordingly
+        for (key_date, list_value) in date_dump.items():
+            news_dump = {}
+            for index, half_link in enumerate(list_value):
+                category = half_link.split('/')[1]
+                print ("*************** Category : {}, #: {} ********************".format(category, (index + 1)))
+                url = ''
                 
-            # Some of the category has no content 
-            # like photo_feature or video
-            if body_content:
-                for body in body_content.findAll('p'):
-                    # check if it the body is empty
-                    # exclude the javascript inside <p></p> tag
-                    # exclude the duplicates from appending
-                    if str(body.text.encode('ascii', 'ignore'))!="" \
-                                    and 'script' not in str(body) \
-                                    and body.text not in news_body:
-                        news_body += body.text
+                # [1:] because of extra / in half_link
+                url = url.join((self.NEWS_LINK, half_link[1:]))
+                print("Link :", url)
 
-                # Get date
-                split_url = url.split('/')
-                cat_eng = split_url[-5]
-                yy = split_url[-4]
-                mm = split_url[-3]
-                dd = split_url[-2]
-                published_date = mm+dd+yy
+                r = urllib.request.urlopen(url).read()
+                soup = BeautifulSoup(r, 'html.parser')
 
-                result = {
-                    'cat_eng' : cat_eng,
-                    'cat_nep' : category,
-                    'eng_date' : published_date,
-                    'nep_date' : nep_date,
-                    'author': author,
-                    'title' : news_title,
-                    'text' : news_body,
-                    'url' : url
-                }
+                title_source = soup.find_all('div', {'class': ['article-header']})
+                if title_source[0].find({'h1', 'h2'}) is not None:
+                    news_title = title_source[0].find({'h1', 'h2'}).text
 
-                news_dump[str(index)] = result
+                author = soup.find('span',{'class': ['author']}).text
+                nep_date = soup.find('time',{'style': ['display:inline-block']}).text           
 
-        return news_dump
+                # Get only the content from 
+                # normal article of main page of each link
+                body_content = soup.find('article', {'class': ['normal']})
+
+                news_body=' '
+
+                # Some of the category has no content 
+                # like photo_feature or video
+                if body_content:
+                    for body in body_content.findAll('p'):
+                        # check if it the body is empty
+                        # exclude the javascript inside <p></p> tag
+                        # exclude the duplicates from appending
+                        if str(body.text.encode('ascii', 'ignore'))!="" \
+                                        and 'script' not in str(body) \
+                                        and body.text not in news_body:
+                            news_body += body.text
+
+                    # To remove unnecessary end of the sentence
+                    # For example - प्रकाशित : वैशाख ३, २०७७ ०८:२७
+                    news_body = ' '.join(news_body.split()[:-6])
+
+                    # Get date
+                    split_url = url.split('/')
+                    cat_eng = split_url[-5]
+                    yy = split_url[-4]
+                    mm = split_url[-3]
+                    dd = split_url[-2]
+                    self.published_date = mm+dd+yy
+
+                    result = {
+                        'cat_eng' : cat_eng,
+                        'cat_nep' : category,
+                        'eng_date' : self.published_date,
+                        'nep_date' : nep_date,
+                        'author': author,
+                        'title' : news_title,
+                        'text' : news_body,
+                        'url' : url
+                    }
+
+                    news_dump[str(index)] = result
             
+            if len(news_dump) > 0:
+                self.dump['category'] = news_dump
+                print("Saving JSON file !!!")
+                self.saveJson(directory=category, input_file=key_date)                 
+                
 
 def main():
-    news_link = sys.argv[1]
-    news_source_name = sys.argv[2]
-    scrappy = Scrapper(news_link=news_link, source=news_source_name)
+    parser = argparse.ArgumentParser("Kantipur Scrapper")
+    parser.add_argument("-n", "--news_link", 
+                        default="https://ekantipur.com/", 
+                        metavar="LINK", help="News Link")
+    parser.add_argument("-s", "--news_source_name", 
+                        default="kantipur", 
+                        metavar="SOURCE", help="News source name")
+    parser.add_argument("-d", "--date", default="",
+                        metavar="DATE", help="Date")
+    
+    args = parser.parse_args()
+    
+    news_link = args.news_link
+    news_source_name = args.news_source_name
+    date = args.date
+    
+    scrappy = Scrapper(news_link=news_link, source=news_source_name, given_date=date)
     scrappy.extractContent()
-    scrappy.saveJson()
 
 if __name__ == '__main__':
     main()
